feat:  retornar 404 quando o post não existir

+++ src/app/controllers/LikeController.ts
@@ -1,36 +1,51 @@
 import { NextFunction, Response } from "express";
 import { LikeService } from "../services/Like.service";
 import { ProtectedRequest } from "../interfaces/protectedRequest";
+import { PostService } from "../services/Post.service";
+import { NotFoundError } from "../../utils/errors/httpErrors";
 
 export class LikeController {
-  constructor(private likeService: LikeService) {}
+  constructor(
+    private likeService: LikeService,
+    private postService: PostService
+  ) {}
 
   public async createPostLike(
     req: ProtectedRequest,
     res: Response,
     next: NextFunction
   ) {
-    const { postId } = req.params;
-    const { userId } = req;
+    try {
+      const { postId } = req.params;
+      const { userId } = req;
 
-    const userLiked = await this.likeService.userLiked({
-      post_id: postId,
-      user_id: userId!,
-    });
+      const postExists = await this.postService.getPostById(postId);
 
-    if (userLiked) {
-      await this.likeService.deletePostLike({
+      if (!postExists) {
+        throw new NotFoundError("Publicação não encontrada");
+      }
+
+      const userLiked = await this.likeService.userLiked({
         post_id: postId,
         user_id: userId!,
       });
 
-      return res.status(204).json();
-    }
-    const postLike = await this.likeService.createPostLike({
-      post_id: postId,
-      user_id: userId!,
-    });
+      if (userLiked) {
+        await this.likeService.deletePostLike({
+          post_id: postId,
+          user_id: userId!,
+        });
 
-    return res.status(201).json(postLike);
+        return res.status(204).json();
+      }
+      const postLike = await this.likeService.createPostLike({
+        post_id: postId,
+        user_id: userId!,
+      });
+
+      return res.status(201).json(postLike);
+    } catch (err) {
+      next(err);
+    }
   }
 }
